Add unit test & code cleanup
