+++ src/components/toolbox/icon/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useTheme } from '../../../utils/theme';
